requirements.txt
Flask==2.3.3
flask-cors==4.0.0
PyMuPDF==1.23.5
faiss-cpu==1.7.4
numpy==1.24.3
scikit-learn==1.3.0
requests==2.31.0
langchain==0.0.284
langchain-community==0.0.5
gunicorn==21.2.0